+++ ImageQ/processor/base.py
@@ -1,5 +1,6 @@
 """@desc 
-        Base Architecture, Request Handlers and Logger for Predictors
+        Provides base architectures to be extended:
+        BasePredictor, BaseSearch and BaseHandler
 
      @author 
          Domnan Diretnan
@@ -19,18 +20,135 @@
  """
 import abc
 import uuid
-from django.core import files
+from abc import ABCMeta, abstractmethod
 from io import BytesIO
-import requests
+from urllib.parse import urlparse
+
 import numpy as np
-from ImageQ.processor.consts import FS, File, IMAGE_TYPES
+import requests
+from bs4 import BeautifulSoup
+
+from django.core import files
+from ImageQ.processor.consts import FS, IMAGE_TYPES, SEARCH_QUERY, File
 from ImageQ.search.models import Prediction
 
 
-
 __all__ = [
-    'BasePredictor', 'RequestHandler'
+    'SearchBase',
+    'BasePredictor', 
+    'RequestHandler', 'UploadHandler',
 ]
+
+class BaseSearch(object):
+
+    __metaclass__ = ABCMeta
+
+    """
+    Search base to be extended by search parsers
+    Every subclass must have two methods `search` amd `parse_single_result`
+    """
+
+    @abstractmethod
+    def search(self, query, page=1):
+        """
+        Master method coordinating search parsing
+        """
+        raise NotImplementedError("subclasses must define method <search>")
+
+    @abstractmethod
+    def parse_single_result(self, single_result):
+        """
+        Every div/span containing a result is passed here to retrieve
+        `title`, `link` and `descr`
+        """
+        raise NotImplementedError("subclasses must define method <parse_results>")
+    
+    def parse_result(self, results):
+        """
+        Runs every entry on the page through parse_single_result
+
+        :param results: Result of main search to extract individual results
+        :type results: list[`bs4.element.ResultSet`]
+        :returns: dictionary. Containing titles, links, netlocs and descriptions.
+        :rtype: dict
+        """
+        titles = []
+        links = []
+        netlocs = []
+        descs = []
+        for each in results:
+            title=link=desc=netloc = " "
+            try:
+                title, link, desc = self.parse_single_result(each)
+                netloc = urlparse(link).netloc
+                ''' Append links and text to a list '''
+                titles.append(title)
+                links.append(link)
+                netlocs.append(netloc)
+                descs.append(desc)
+            except Exception as e:
+                print(e)
+        search_results = {'titles': titles,
+                          'links': links,
+                          'netlocs': netlocs,
+                          'descriptions': descs}
+        return search_results
+    
+    @staticmethod
+    def parse_query(query):
+        """
+        Replace spaces in query
+
+        :param query: query to be processed
+        :type query: str
+        :rtype: str
+        """
+        return query.replace(" ", "%20")
+    
+    @staticmethod
+    def getSource(url):
+        """
+        Returns the source code of a webpage.
+
+        :rtype: string
+        :param url: URL to pull it's source code
+        :return: html source code of a given URL.
+        """
+        import requests
+        # headers = {'User-Agent': 'Mozilla/5.0 (Windows NT 6.1; WOW64; rv:20.0) Gecko/20100101 Firefox/20.0'}
+        # prevent caching
+        headers={'Cache-Control': 'no-cache'}
+        try:
+            response = requests.get(url, headers=headers)
+            html = response.text
+        except Exception as e:
+            raise Exception('ERROR: {}\n'.format(e))
+        return str(html)   
+
+    @staticmethod
+    def get_soup(raw_query, engine="Google", page=1):
+        """
+        Get the html soup of a query
+
+        :param raw_query: unprocessed query string
+        :type raw_query: str
+        :param engine: search engine to make use of, defaults to google
+        :type engine: str
+        :param page: page to return
+        :type page: int
+        :rtype: `bs4.element.ResultSet`
+        """
+        # replace spaces in string
+        query = BaseSearch.parse_query(raw_query)
+        search_fmt_string = SEARCH_QUERY[engine]
+        if engine=="Google":
+            search_url = search_fmt_string.format(query, page)
+        if engine=="Yahoo":
+            offset = (page * 10) - 9
+            search_url = search_fmt_string.format(query, offset)
+        html = BaseSearch.getSource(search_url)
+        return BeautifulSoup(html, 'lxml')
+
 
 class BasePredictor(object):
     """This is the basic predictor extended by all other predictors
@@ -63,50 +181,18 @@ class BasePredictor(object):
             raise AttributeError("Attribute <prediction_api> is of type None")
         return r.content
 
-class RequestHandler:
-    """Request handler
-    """
-    def __init__(self, image):
-        """Constructor
+class BaseHandler(object):
+    __metaclass__ = ABCMeta
 
-        :param image:  A dictionary containing image data including it's url, type
-        :type image: dict
-        :param image_name: final name of image when downloaded
-        :type image_name: str
-        """
-        self.image = image
-        response = requests.get(image.get('url'))
-        if response.status_code != requests.codes.ok:
-            return Exception("Something just happened Right Now")
-        self.fp = BytesIO()
-        self.fp.write(response.content)
-
-    def save(self):
-        """Save the image in the Prediction Model
-        """
-        # Generate a random string as file_name
-        image_name = uuid.uuid1().hex
-        prediction = Prediction()
-        # Save the Image without Downloading it
-        prediction.image.save(f'{image_name}.{self.image.get("ext")}', files.File(self.fp))
-        return prediction
-
-class UploadHandler:
-    """
-    Uploaded Image Handler
-
-    :param image: image to initialize class with
-    :tyoe image: `django.core.files.uploadedfile.InMemoryUploadedFile`
-    """
-    def __init__(self, image):
-        if image:
-            self.image = image
-        else:
-            raise AttributeError(f"cannot init class with image of type {type(image)}")
-        if self.image.content_type.startswith("image/"):
-            self.ext = self.image.content_type[7:]
-        else:
-            raise ValueError("content-type <image/*> expected")
+    @property
+    @abstractmethod
+    def image_data(self):
+        return NotImplementedError("self.image_data must be defined")
+    
+    @property
+    @abstractmethod
+    def ext(self):
+        return NotImplementedError("self.ext must be defined")
 
     def save(self):
         """Save the image in the Prediction Model
@@ -114,5 +200,5 @@ class UploadHandler:
         image_name = uuid.uuid1().hex
         prediction = Prediction()
         # Save the Image without Downloading it
-        prediction.image.save(f'{image_name}.{self.ext}', self.image)
+        prediction.image.save(f'{image_name}.{self.ext}', self.image_data)
         return prediction
